+++ contracts/src/IncredibleSquaringTaskManager.sol
@@ -81,7 +81,7 @@ contract IncredibleSquaringTaskManager is
     /* FUNCTIONS */
     // NOTE: this function creates new task, assigns it a taskId
     function createNewTask(
-        uint256 dolarDatetime,
+        uint32 dolarDatetime,
         uint32 quorumThresholdPercentage,
         bytes calldata quorumNumbers
     ) external onlyTaskGenerator {
